+++ quantforge/__init__.py
@@ -256,7 +256,7 @@ from .callable_bond import (
 from .convertible_lattice import convertible_bond_lattice
 from .shout import shout_call, ladder_call
 from .installment import installment_call
-from .double_barrier import double_knockout_call
+from .double_barrier import double_knockout_call, double_knockin_call
 from .range_accrual import range_accrual_note
 from .evt import (
     hill_estimator, gpd_fit_pot, gpd_var, gpd_expected_shortfall,
@@ -600,7 +600,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "1.539.0"
+__version__ = "1.540.0"
 
 __all__ = [
     "OptionType",
@@ -1003,6 +1003,7 @@ __all__ = [
     "ladder_call",
     "installment_call",
     "double_knockout_call",
+    "double_knockin_call",
     "range_accrual_note",
     "hill_estimator",
     "gpd_fit_pot",